Moved from NSString to NSURL on Source
In 10.7, the source file is really much better being an NSURL and NOT an
NSString. This is because of the way the OS is now handling the Open and
Save dialogs as well as the file handling in general. Because of this, it
makes a lot of sense to move to the NSURL, and that meant that we needed to
change a few signatures, and then change the way we're creating the filenames.
It's not a massive change, but it's a goot chunk of work, and we should be
good to go for a while with this scheme.