Added drawing of results
We needed to start somewhere, and plotting the Voltage was a good first
step. This does so by saving the voltage - as a fraction [0..1] and then
map that between the two colors on the plot at the drawing time. It
works, and now points to issues we have with the data - which were once
all cleared up, but now we have an issue. That's next.