Fixed up the Simulation
We had a classic problem - the latest macOS 10.14 doesn't clear out the
memory created with malloc - and it shouldn't have to. So we used the
vDSP funcitons in the Accelerate framework to quickly fill the malloc-ed
arrays with zeros, and then everything was stable. It was fantastic. We
also added in a little timer, and the speed is amazing.