Switch to Multi-Color Graphing
I really wanted to have multiple colors in the graphing, and it came to
me how to do it pretty easily. By placing the colors in a simple array,
and then looking at the value to interpolate, we can find the two colors
that we need to interpolate between. Not bad at all.