Added the Electric Field to the answers
We had a point-by-point calculation of the Electric Field - magnitude
and direction, but now we calculate it for all solutions. This then
makes it much easier to plot the components of the electric field, and
that's a step in the right direction. The next thing, we need to add the
isobars for the plot.