Added Switching V(x,y) and E(x,y)
The next little addition is the ability to select what you wanted to see
plotted when the simulation is run. I put this in the menu so that it
was easy to wire up to keyboard shortcuts, and at the same time, give
the user a visual reminder of what they are seeing - V or E. This works
like a charm, and flips back and forth as good as you please.