Added metrics on the drawing
It seemed like a good idea to see how long the drawing was taking, and
it turns out to be really informative. We are taking more than 3x the
time to graph the solution - versus solve it. At the same time, we added
blue on the very low end, to see how it looks, and I like the effect, it
makes the low-end look "cool", and that makes more visual sense to me.