Added the Inventory to graph
The next big thing was to add in all the items under simulation to the
graph, so that the user could see where they were, and then the solution
would make so much more sense. This was another interesting experience
with Quartz 2D, and turned out very nicely. At the same time, I realized
that I had the location of the rectangle all wrong in the set-up, and
needed to correct that as well. Now everything matches up very nicely.