Added some setters and getters
I was looking through the code, and realized I really needed to add in a
few setters and getters for the ResultsView - so that it was a little
nicer than dealing with instance variables all the time. Minor point,
and as I was in the code, I was looking for places to use the vDSP
functions to speed things up, but there just weren't enough simple
vector operations - even the min and max calculations were masked, so in
the end, this is not something I can take advantage of yet.