Cleaned up the Drawing Code
We really had a big mess on the drawing code - everything was in the one
method, and it was not as clean as it could be. Now we've added the
methods to make the line count the same, but we have added the ability
to compose things a little differently. This is a very good thing. At
the same time, we have cleared up all the RGB work for the drawing, and
used the NSColor methods to set the stroke or fill color directly. This
makes the code a lot simpler. Much better.