Added Contour Lines to Graph
With the refactoring, it was time to get down to the Contour Lines, and
see if I could get them working. Turns out, it wasn't as hard as I had
feared, but then I figured it out on the walk into work this morning.
What a slick answer to this problem. Love it.