Cleaned up whitespace on rects
When drawing a lot of the color blocks, we end up with very thin white
lines between the blocks, and that's fine - but we really don't want to
see that. So let's add a little bit to the width and height of the
blocks so that they overlap just a bit, and the way they are drawn makes
sure that the most accurate colors are represented. This looks a lot
better now.