Respect the Workspace Ratio
We were filling up the width and height of the drawing surface -
regardless of the stated ratio of the sides of the Workspace. This
change respects the ratio of the workspace - so a square ALWAYS looks
like a square - and there is just white space on those sides of the
display where the drawing can't fit without distortion. This is nice in
that it makes sure that what's requested is presented in the same way.