Fixed up the registration on the graph
We had a problem in how we were displaying the data - the colored
rectangles were nice, but we couldn't extend the contours to cover all
the boxes - because a contour takes 4 pts. It makes more sense to have
the plotting data function in a similar manner - where 4 pts work
together to create the value for the colored rectangle. In this case, we
have chosen a simple MAX() function - and it works just fine.