Added the direction of the E Field
For the display of the electric field, we kept the color for the
magnitude, but took out the contours as those really aren't what you
want to see on an electric field map - you want to see the *direction*
of the field, and now that's what we have. We added in a simple arrow,
and then used CGAffineTransforms to rotate it per the direction of the
field, scale it, and translate it into the view. All this was pretty
slick, and it's ready to go.